+++ pobject/pobject.Core/Transactions/ITransactionsService.cs
@@ -13,7 +13,7 @@ namespace pobject.Core.Transactions
 
         public StoreCode deposit(Transaction_Deposit request,string _bearer_token);
 
-        public int updateamount(string Referral_UserId, int amount, string email);
+        public int updateamount(string Referral_UserId, float amount, string email);
         public StoreCode sentamount(SetAmount request, string email);
     }
 }
